split PdfContainer from PdfView for refactoring
